module-2-nemo-agent-toolkit/deep-dive/pyproject.toml
[project]
name = "nat_email_phishing_analyzer"
dynamic = ["version"]
dependencies = [
  "nvidia-nat[all]>=1.4.1",
  "langchain<1.0.0",
  "jupyterlab"
]
requires-python = ">=3.12"
description = "Email Phishing Analyzer - NeMo Agent Toolkit Deep Dive"
keywords = ["ai", "rag", "agents"]
classifiers = ["Programming Language :: Python"]

[project.entry-points.'nat.components']
nat_email_phishing_analyzer = "nat_email_phishing_analyzer.register"
